+++ CRM/Contact/DataProcessorContactSearch.php
@@ -55,6 +55,8 @@ class CRM_Contact_DataProcessorContactSearch implements UIFormOutputInterface {
 
     $form->add('wysiwyg', 'help_text', E::ts('Help text for this search'), array('rows' => 6, 'cols' => 80));
 
+    $form->add('text', 'no_result_text', E::ts('No result text'), array('class' => 'huge'), false);
+
     $form->add('checkbox', 'expanded_search', E::ts('Expand criteria form initially'));
 
     // navigation field
@@ -74,6 +76,11 @@ class CRM_Contact_DataProcessorContactSearch implements UIFormOutputInterface {
       if (isset($output['configuration']) && is_array($output['configuration'])) {
         if (isset($output['configuration']['contact_id_field'])) {
           $defaults['contact_id_field'] = $output['configuration']['contact_id_field'];
+        }
+        if (isset($output['configuration']['no_result_text'])) {
+          $defaults['no_result_text'] = $output['configuration']['no_result_text'];
+        } else {
+          $defaults['no_result_text'] = E::ts('No results');
         }
         if (isset($output['configuration']['navigation_id'])) {
           $defaults['navigation_parent_path'] = $navigation->getNavigationParentPathById($output['configuration']['navigation_id']);
@@ -121,6 +128,7 @@ class CRM_Contact_DataProcessorContactSearch implements UIFormOutputInterface {
     $output['permission'] = $submittedValues['permission'];
     $configuration['contact_id_field'] = $submittedValues['contact_id_field'];
     $configuration['hidden_fields'] = $submittedValues['hidden_fields'];
+    $configuration['no_result_text'] = $submittedValues['no_result_text'];
     $configuration['navigation_parent_path'] = $submittedValues['navigation_parent_path'];
     $configuration['hide_id_field'] = $submittedValues['hide_id_field'];
     $configuration['help_text'] = $submittedValues['help_text'];