+++ CRM/DataprocessorSearch/ActivitySearch.php
@@ -56,6 +56,7 @@ class CRM_DataprocessorSearch_ActivitySearch implements UIFormOutputInterface {
     ));
 
     $form->add('wysiwyg', 'help_text', E::ts('Help text for this search'), array('rows' => 6, 'cols' => 80));
+    $form->add('text', 'no_result_text', E::ts('No result text'), array('class' => 'huge'), false);
     $form->add('checkbox', 'expanded_search', E::ts('Expand criteria form initially'));
 
     // navigation field
@@ -83,6 +84,11 @@ class CRM_DataprocessorSearch_ActivitySearch implements UIFormOutputInterface {
         }
         if (isset($output['configuration']['hidden_fields'])) {
           $defaults['hidden_fields'] = $output['configuration']['hidden_fields'];
+        }
+        if (isset($output['configuration']['no_result_text'])) {
+          $defaults['no_result_text'] = $output['configuration']['no_result_text'];
+        } else {
+          $defaults['no_result_text'] = E::ts('No results');
         }
         if (isset($output['configuration']['help_text'])) {
           $defaults['help_text'] = $output['configuration']['help_text'];
@@ -121,6 +127,7 @@ class CRM_DataprocessorSearch_ActivitySearch implements UIFormOutputInterface {
     $configuration['activity_id_field'] = $submittedValues['activity_id_field'];
     $configuration['navigation_parent_path'] = $submittedValues['navigation_parent_path'];
     $configuration['hide_id_field'] = $submittedValues['hide_id_field'];
+    $configuration['no_result_text'] = $submittedValues['no_result_text'];
     $configuration['hidden_fields'] = $submittedValues['hidden_fields'];
     $configuration['help_text'] = $submittedValues['help_text'];
     $configuration['expanded_search'] = isset($submittedValues['expanded_search']) ? $submittedValues['expanded_search'] : false;