+++ CRM/DataprocessorSearch/CaseSearch.php
@@ -61,6 +61,7 @@ class CRM_DataprocessorSearch_CaseSearch implements UIFormOutputInterface {
     ));
 
     $form->add('wysiwyg', 'help_text', E::ts('Help text for this search'), array('rows' => 6, 'cols' => 80));
+    $form->add('text', 'no_result_text', E::ts('No result text'), array('class' => 'huge'), false);
     $form->add('checkbox', 'expanded_search', E::ts('Expand criteria form initially'));
 
     // navigation field
@@ -95,6 +96,11 @@ class CRM_DataprocessorSearch_CaseSearch implements UIFormOutputInterface {
         if (isset($output['configuration']['help_text'])) {
           $defaults['help_text'] = $output['configuration']['help_text'];
         }
+        if (isset($output['configuration']['no_result_text'])) {
+          $defaults['no_result_text'] = $output['configuration']['no_result_text'];
+        } else {
+          $defaults['no_result_text'] = E::ts('No results');
+        }
         if (isset($output['configuration']['expanded_search'])) {
           $defaults['expanded_search'] = $output['configuration']['expanded_search'];
         }
@@ -128,6 +134,7 @@ class CRM_DataprocessorSearch_CaseSearch implements UIFormOutputInterface {
     $output['permission'] = $submittedValues['permission'];
     $configuration['contact_id_field'] = $submittedValues['contact_id_field'];
     $configuration['case_id_field'] = $submittedValues['case_id_field'];
+    $configuration['no_result_text'] = $submittedValues['no_result_text'];
     $configuration['hidden_fields'] = $submittedValues['hidden_fields'];
     $configuration['navigation_parent_path'] = $submittedValues['navigation_parent_path'];
     $configuration['hide_id_fields'] = $submittedValues['hide_id_fields'];