+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Added current user to contact filter.
 * Added data source for permissioned contact (#25).
 * Fixed issue with configuration contact source sub type filter.
+* Added a no result text to the outputs.
 
 # Version 1.3.0